+++ src/AbstractMoney.php
@@ -14,6 +14,9 @@ use JsonSerializable;
 
 /**
  * Base class for Money and RationalMoney.
+ *
+ * Please consider this class sealed: extending this class yourself is not supported, and breaking changes (such as
+ * adding new abstract methods) can happen at any time, even in a minor version.
  */
 abstract class AbstractMoney implements MoneyContainer, JsonSerializable
 {